Try rootless travis w/o docker

+++ tests/ttfautohint/ttfautohint.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
-
-exec docker run --rm -a STDIN -a STDOUT -a STDERR -i ttfautohint:1.5 "$@"
+if command -v docker >/dev/null 2>&1; then
+	exec docker run --rm -a STDIN -a STDOUT -a STDERR -i ttfautohint:1.5 "$@"
+else
+	exec ttfautohint "$@"
+fi
